Move logic to data.ts
Moved logic from generic-table.component.ts to data.ts.
This allows for sorting and filtering.
There are still problems with initial display.
