+++ app/components/output/generic-table/generic-table.component.ts
@@ -16,7 +16,6 @@ import { DomSanitizer, SafeHtml } from "@angular/platform-browser";
 import { isNullOrUndefined } from "util";
 import { Http } from "@angular/http";
 declare let anchors;
-declare let moment: any;
 
 @Component({
     selector: 'generictable',
@@ -48,8 +47,6 @@ export class GenericTableComponent implements AfterViewChecked, OnChanges {
     @Input() orderOption: Array<number> = new Array<number>();
     @Output() orderOptionChange: EventEmitter<any> = new EventEmitter();
 
-    @Output() repoLabels: {[column: string]: Array<{htmlChilds: string, content: string}>} = {};
-
     private ctrlCounter: number = 0;
 
     constructor(private ar: ApplicationRef,
@@ -57,77 +54,6 @@ export class GenericTableComponent implements AfterViewChecked, OnChanges {
                 private sanitization: DomSanitizer,
                 private http: Http,
                 private cd: ChangeDetectorRef) {
-    }
-
-    private getRepoLabels(td: TableData, data: Data) {
-        if (isNullOrUndefined(data.properties["Repo"]) || isNullOrUndefined(data.properties["Repo"].list[0])) {
-            return [];
-        }
-        if (!isNullOrUndefined(this.repoLabels[data.tag])) {
-            return this.repoLabels[data.tag];
-        }
-        const repoUrl: string = data.properties["Repo"].list[0].content;
-        const url = this.repoQueryBuildUrl(repoUrl);
-        if (url === "") {
-            return [];
-        }
-        this.http.get(url).toPromise().then(res => {
-            const d: any = {};
-            const commitDate = new Date(res.json()[0].commit.author.date);
-            let child: string = "The last commit is ";
-            const cd = moment(commitDate);
-            const now = moment();
-            const dateStrings = ['years', 'months', 'days', 'hours', 'minutes', 'seconds'];
-            for (const s of dateStrings) {
-                const diff = Math.abs(now.diff(cd, s));
-                if (diff !== 0) {
-                    child += diff;
-                    // append unit in singular or plural
-                    child += " " + (diff === 1 ? s.substr(0, s.length - 1) : s);
-                    break;
-                }
-            }
-            child += " old";
-            d.htmlChilds = child;
-            for (const value in td.values) {
-                if (isNullOrUndefined(td.values[value]["min-age"])) {
-                    continue;
-                }
-
-                const min = td.values[value]["min-age"];
-                const minUnit = td.values[value]["min-age-unit"];
-                const minDiff = Math.abs(now.diff(cd, minUnit));
-
-                const max = td.values[value]["max-age"];
-                const maxUnit = td.values[value]["max-age-unit"];
-                const maxDiff = Math.abs(now.diff(cd, maxUnit));
-
-                // min === -1 => no limit
-                // same for max
-                if ((min === -1 || min <= minDiff) && (max === -1 || max > maxDiff)) {
-                    d.content = value;
-                    if (isNullOrUndefined(this.repoLabels[data.tag])) {
-                        this.repoLabels[data.tag] = [];
-                    }
-                    for (const item of this.repoLabels[data.tag]) {
-                        if (item.content === d.content) {
-                            this.repoLabels[data.tag].splice(this.repoLabels[data.tag].indexOf(item), 1);
-                        }
-                    }
-                    this.repoLabels[data.tag].push(d);
-                }
-            }
-            this.cd.markForCheck();
-        });
-    }
-
-    private repoQueryBuildUrl(repoUrl: string) {
-        let url: string;
-        if (/https?:\/\/github\.com.*/.test(repoUrl)) {
-            url = repoUrl.replace(/https?:\/\/github.com/, "https://api.github.com/repos");
-            url += url.endsWith("/") ? "commits" : "/commits";
-        }
-        return url;
     }
 
     private orderClick(e: MouseEvent, value: string) {
